+++ test/modules/fs-test.js
@@ -2,7 +2,7 @@ describe('modules/fs', function() {
   var fs = window.xpcModule.require('fs');
 
   describe('dir', function() {
-    var fileName = __dirname + '../fixtures/test-dir/';
+    var fileName = __dirname + '/../fixtures/test-dir/';
 
     function cleanup() {
       if (fs.existsSync(fileName))
@@ -16,6 +16,11 @@ describe('modules/fs', function() {
       expect(fs.existsSync(fileName)).to.be(false);
       fs.mkdirSync(fileName, 0655);
       expect(fs.existsSync(fileName)).to.be(true);
+
+      var stat = fs.statSync(fileName);
+      expect(stat.isDirectory()).to.be(true);
+      expect(stat.isFile()).to.be(false);
+
 
       fs.rmdirSync(fileName);
       expect(fs.existsSync(fileName)).to.be(false);
@@ -55,6 +60,14 @@ describe('modules/fs', function() {
   describe('read', function() {
     var fileName = __dirname + '/../fixtures/file.js';
     var expected = 'local file';
+
+    describe('statSync', function() {
+      it('should have stat functions', function() {
+        var stat = fs.statSync(fileName);
+        expect(stat.isDirectory()).to.be(false);
+        expect(stat.isFile()).to.be(true);
+      });
+    });
 
     describe('readFileSync', function() {
       it('should be able to read files', function() {